Add pointblank.mcp package and rename MCP server
