+++ pointblank/mcp/__init__.py
@@ -1,0 +1,3 @@
+from pointblank.mcp.server import mcp
+
+__all__ = ["mcp"]
