+++ pointblank/mcp/__main__.py
@@ -1,0 +1,4 @@
+from pointblank.mcp.server import mcp
+
+if __name__ == "__main__":
+    mcp.run(transport="stdio")
