+++ tests/test_mcp_server.py
@@ -4,7 +4,7 @@ import pandas as pd
 import pytest
 
 from fastmcp import Client
-from pointblank_mcp_server.pointblank_server import mcp
+from pointblank.mcp.server import mcp
 
 
 @pytest.fixture(scope="module")
