+++ user_guide/08-mcp-server/01-mcp-quick-start.qmd
@@ -28,7 +28,7 @@ pip install pointblank[mcp]
   "servers": {
     "pointblank": {
       "command": "python",
-      "args": ["-m", "pointblank_mcp_server.pointblank_server"]
+      "args": ["-m", "pointblank.mcp"]
     }
   }
 }
